Add new signals and methods to chad skin...
To deal with attacks and combos. Now, with animations, attacks
should accept input only for some frames and if no input
is given, it should go back to idle. This new methods and signals
are to support this logic, and be integrated to the Character Actions
later, and maybe to the base skin?