Add Knockback Struct and move some enums to CyclycErrorHelper
Wile implementing Launch I noticed that I need to pass launch data to
hurt as well, as any hit on air restarts the knockout state, but the hurtbox
shouldn't need to know any of it.

So by creating a Knockback Struct I can make the Hurtbox work in a more
generic way, and the Knockback State deal with the struct it will
receive, but this generated a lot of CyclicErrors, so I had to move the
enums.