Add most assets for tax man area and build it on stage 01
At this point I have stage 01 mostly complete, at least we have all
the areas in. Doing the tax man area I noticed some z-sorting problems
and will need to split the tori, and the "brick gate" into the tax man
area to fix them. Also still missing the tax man's doors

But it was good to have this and being able to test walking around in
it, as I also found and fixed some bugs with level collision